requirements.txt
alembic==1.14.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bcrypt==4.2.1
blinker==1.9.0
cffi==1.17.1
click==8.1.8
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
psycopg2==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2025.1
SQLAlchemy==2.0.38
APScheduler==3.11.0
pytest==8.3.4
typing_extensions==4.12.2
Werkzeug==3.1.3
requests==2.31.0
tqdm==4.66.2
